requirements.txt
pytest>=6.1.2
pytest-cov>=2.6.1
pytest-xdist>=2.1.0
setuptools>=40.6.3
twine>=1.12.1
pyyaml>=5.1
